Update README.md
The objective of this assignment is to get you comfortable with numbering systems, bitwise operations, and common binary operations. The programming assignment should be implemented in Java. You are required to work alone on this assignment. The assignment accounts for 7.5% towards your cumulative course grade.

This assignment may be modified to clarify any questions (and the version number incremented), but the crux of the assignment and the distribution of points will not change. Any there any changes to the assignment, all changes will be document in the “Change History” section of this assignment,

1 Assignment Description

The goal of this assignment is for you to learn how to convert between numbering systems and how binary operations work. A secondary goal is for you to continue to develop your programming skills. This program should be developed in Java version 8, and your classes must reside in the package “cs250.hw1”.

Your program will be provided with 3 arguments at the command line. You are required to perform a set of operations over these arguments. These operations have been broken up into tasks: the points distribution for each task and the restrictions (and accompanying deductions) are specified in the grading section of this assignment.